created Path versions of Load Video and Load Images nodes
