+++ videohelpersuite/nodes.py
@@ -1,24 +1,17 @@
 import os
-import hashlib
 import json
 import subprocess
 import shutil
-import re
-import time
 import numpy as np
 from typing import List
-import torch
-from PIL import Image, ImageOps
+from PIL import Image
 from PIL.PngImagePlugin import PngInfo
-import cv2
-from pathlib import Path
 
 import folder_paths
-from comfy.k_diffusion.utils import FolderOfImages
-from comfy.utils import common_upscale
 from .logger import logger
 from .image_latent_nodes import DuplicateImages, DuplicateLatents, GetImageCount, GetLatentCount, MergeImages, MergeLatents, SelectEveryNthImage, SelectEveryNthLatent, SplitLatents, SplitImages
-from .utils import calculate_file_hash, get_sorted_dir_files_from_directory
+from .load_video_nodes import LoadVideoUpload, LoadVideoPath
+from .load_images_nodes import LoadImagesFromDirectoryUpload, LoadImagesFromDirectoryPath
 
 folder_paths.folder_names_and_paths["video_formats"] = (
     [
@@ -223,290 +216,13 @@ class VideoCombine:
         ]
         return {"ui": {"gifs": previews}}
 
-class LoadVideo:
-    @classmethod
-    def INPUT_TYPES(s):
-        video_extensions = ['webm', 'mp4', 'mkv', 'gif']
-        input_dir = folder_paths.get_input_directory()
-        files = []
-        for f in os.listdir(input_dir):
-            if os.path.isfile(os.path.join(input_dir, f)):
-                file_parts = f.split('.')
-                if len(file_parts) > 1 and (file_parts[-1] in video_extensions):
-                    files.append(f)
-        return {"required": {
-                    "video": (sorted(files), {"video_upload": True}),
-                     "force_rate": ("INT", {"default": 0, "min": 0, "max": 24, "step": 1}),
-                     "force_size": (["Disabled", "256x?", "?x256", "256x256", "512x?", "?x512", "512x512"],),
-                     "frame_load_cap": ("INT", {"default": 0, "min": 0, "step": 1}),
-                     "skip_first_frames": ("INT", {"default": 0, "min": 0, "step": 1}),
-                     },}
-
-    CATEGORY = "Video Helper Suite 🎥🅥🅗🅢"
-
-    RETURN_TYPES = ("IMAGE", "INT",)
-    RETURN_NAMES = ("IMAGE", "frame_count",)
-    FUNCTION = "load_video"
-
-    def is_gif(self, filename):
-        file_parts = filename.split('.')
-        return len(file_parts) > 1 and file_parts[-1] == "gif"
-
-    def target_size(self, width, height, force_size):
-        if force_size != "Disabled":
-            force_size = force_size.split("x")
-            if force_size[0] == "?":
-                width = (width*int(force_size[1]))//height
-                #Limit to a multple of 8 for latent conversion
-                #TODO: Consider instead cropping and centering to main aspect ratio
-                width = int(width)+4 & ~7
-                height = int(force_size[1])
-            elif force_size[1] == "?":
-                height = (height*int(force_size[0]))//width
-                height = int(height)+4 & ~7
-                width = int(force_size[0])
-        return (width, height)
-
-    known_exceptions = []
-    def load_video(self, **kwargs):
-        backends = [self.load_video_cv, self.load_video_ffmpeg]
-        if preferred_backend == 'ffmpeg':
-            backends.reverse()
-        for backend in backends:
-            try:
-                return backend(**kwargs)
-            except Exception as e:
-                latest_error = e
-                if e.__str__() not in self.known_exceptions:
-                    self.known_exceptions.append(e.__str__())
-                    logger.warn(f"Load with {backend.__name__} failed due to: {e}")
-        error_string = f"Failed to load video: {kwargs['video']}\ndue to: {latest_error.__str__()}"
-        raise RuntimeError(error_string)
-
-
-    def load_video_cv(self, video, force_rate, force_size, frame_load_cap, skip_first_frames):
-        try:
-            video_cap = cv2.VideoCapture(folder_paths.get_annotated_filepath(video))
-            if not video_cap.isOpened():
-                raise ValueError(f"{video} could not be loaded with cv.")
-            # set video_cap to look at start_index frame
-            images = []
-            total_frame_count = 0
-            frames_added = 0
-            base_frame_time = 1/video_cap.get(cv2.CAP_PROP_FPS)
-            width = video_cap.get(cv2.CAP_PROP_FRAME_WIDTH)
-            height = video_cap.get(cv2.CAP_PROP_FRAME_HEIGHT)
-            if force_rate == 0:
-                target_frame_time = base_frame_time
-            else:
-                target_frame_time = 1/force_rate
-            time_offset=target_frame_time - base_frame_time
-            while video_cap.isOpened():
-                if time_offset < target_frame_time:
-                    is_returned, frame = video_cap.read()
-                    # if didn't return frame, video has ended
-                    if not is_returned:
-                        break
-                    time_offset += base_frame_time
-                if time_offset < target_frame_time:
-                    continue
-                time_offset -= target_frame_time
-                # if not at start_index, skip doing anything with frame
-                total_frame_count += 1
-                if total_frame_count <= skip_first_frames:
-                    continue
-                # TODO: do whatever operations need to happen, like force_size, etc
-
-                # opencv loads images in BGR format (yuck), so need to convert to RGB for ComfyUI use
-                # follow up: can videos ever have an alpha channel?
-                frame = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
-                # convert frame to comfyui's expected format (taken from comfy's load image code)
-                image = Image.fromarray(frame)
-                image = ImageOps.exif_transpose(image)
-                image = np.array(image, dtype=np.float32) / 255.0
-                image = torch.from_numpy(image)[None,]
-                images.append(image)
-                frames_added += 1
-                # if cap exists and we've reached it, stop processing frames
-                if frame_load_cap > 0 and frames_added >= frame_load_cap:
-                    break
-        finally:
-            video_cap.release()
-        if len(images) == 0:
-            raise RuntimeError("No frames generated")
-        images = torch.cat(images, dim=0)
-        if force_size != "Disabled":
-            new_size = self.target_size(width, height, force_size)
-            if new_size[0] != width or new_size[1] != height:
-                s = images.movedim(-1,1)
-                s = common_upscale(s, new_size[0], new_size[1], "lanczos", "disabled")
-                images = s.movedim(1,-1)
-        # TODO: raise an error maybe if no frames were loaded?
-        return (images, frames_added)
-
-    def load_video_ffmpeg(self, video, force_rate, force_size, frame_load_cap, skip_first_frames):
-        if ffmpeg_path is None:
-            return self.load_video_cv_fallback(video, force_rate, force_size, frame_load_cap, skip_first_frames)
-        # otherwise, continue with ffmpeg
-        video_path = folder_paths.get_annotated_filepath(video)
-        args_dummy = [ffmpeg_path, "-i", video_path, "-f", "null", "-"]
-        with subprocess.Popen(args_dummy, stdout=subprocess.DEVNULL, stderr=subprocess.PIPE) as proc:
-            for line in proc.stderr.readlines():
-                match = re.search(", ([1-9]|\\d{2,})x(\\d+)",line.decode('utf-8'))
-                if match is not None:
-                    size = [int(match.group(1)), int(match.group(2))]
-                    break
-        args_all_frames = [ffmpeg_path, "-an", "-i", video_path, "-v", "error",
-                             "-pix_fmt", "rgb24", "-vsync", "2"]
-
-        vfilters = []
-        if force_rate != 0:
-            vfilters.append("fps=fps="+str(force_rate) + ":round=up")
-        if skip_first_frames > 0:
-            vfilters.append(f"select=gt(n\\,{skip_first_frames-1})")
-        if force_size != "Disabled":
-            size = self.target_size(size[0], size[1], force_size)
-            vfilters.append(f"scale={size[0]}:{size[1]}")
-        if len(vfilters) > 0:
-            args_all_frames += ["-vf", ",".join(vfilters)]
-        if frame_load_cap > 0:
-            args_all_frames += ["-frames:v", str(frame_load_cap)]
-
-        args_all_frames += ["-f", "rawvideo", "-"]
-        images = []
-        with subprocess.Popen(args_all_frames, stdout=subprocess.PIPE) as proc:
-            #Manually buffer enough bytes for an image
-            bpi = size[0]*size[1]*3
-            current_bytes = bytearray(bpi)
-            current_offset=0
-            while True:
-                bytes_read = proc.stdout.read(bpi - current_offset)
-                if bytes_read is None:#sleep to wait for more data
-                    time.sleep(.2)
-                    continue
-                if len(bytes_read) == 0:#EOF
-                    break
-                current_bytes[current_offset:len(bytes_read)] = bytes_read
-                current_offset+=len(bytes_read)
-                if current_offset == bpi:
-                    images.append(np.array(current_bytes, dtype=np.float32).reshape(size[1], size[0], 3) / 255.0)
-                    current_offset = 0
-
-        if len(images) == 0:
-            raise RuntimeError("No frames generated")
-        images = torch.from_numpy(np.stack(images))
-        return (images, images.size(0))
-
-    @classmethod
-    def IS_CHANGED(s, video, **kwargs):
-        image_path = folder_paths.get_annotated_filepath(video)
-        m = hashlib.sha256()
-        with open(image_path, 'rb') as f:
-            m.update(f.read())
-        return m.digest().hex()
-
-    @classmethod
-    def VALIDATE_INPUTS(s, video, **kwargs):
-        if not folder_paths.exists_annotated_filepath(video):
-            return "Invalid image file: {}".format(video)
-
-        return True
-
-
-class LoadImagesFromDirectory:
-    @classmethod
-    def INPUT_TYPES(s):
-        input_dir = folder_paths.get_input_directory()
-        directories = []
-        for item in os.listdir(input_dir):
-            if not os.path.isfile(os.path.join(input_dir, item)) and item != "clipspace":
-                directories.append(item)
-        return {
-            "required": {
-                "directory": (directories,),
-            },
-            "optional": {
-                "image_load_cap": ("INT", {"default": 0, "min": 0, "step": 1}),
-                "skip_first_images": ("INT", {"default": 0, "min": 0, "step": 1}),
-                "select_every_nth": ("INT", {"default": 1, "min": 1, "step": 1}),
-            }
-        }
-    
-    RETURN_TYPES = ("IMAGE", "MASK", "INT")
-    FUNCTION = "load_images"
-
-    CATEGORY = "Video Helper Suite 🎥🅥🅗🅢"
-
-    def load_images(self, directory: str, image_load_cap: int = 0, skip_first_images: int = 0, select_every_nth: int = 1):
-        directory = folder_paths.get_annotated_filepath(directory.strip())
-        if not os.path.isdir(directory):
-            raise FileNotFoundError(f"Directory '{directory} cannot be found.")
-
-        dir_files = get_sorted_dir_files_from_directory(directory, skip_first_images, select_every_nth, FolderOfImages.IMG_EXTENSIONS)
-
-        if len(dir_files) == 0:
-            raise FileNotFoundError(f"No files in directory '{directory}'.")
-
-        images = []
-        masks = []
-
-        limit_images = False
-        if image_load_cap > 0:
-            limit_images = True
-        image_count = 0
-
-        for image_path in dir_files:
-            if limit_images and image_count >= image_load_cap:
-                break
-            i = Image.open(image_path)
-            i = ImageOps.exif_transpose(i)
-            image = i.convert("RGB")
-            image = np.array(image).astype(np.float32) / 255.0
-            image = torch.from_numpy(image)[None,]
-            if 'A' in i.getbands():
-                mask = np.array(i.getchannel('A')).astype(np.float32) / 255.0
-                mask = 1. - torch.from_numpy(mask)
-            else:
-                mask = torch.zeros((64,64), dtype=torch.float32, device="cpu")
-            images.append(image)
-            masks.append(mask)
-            image_count += 1
-        
-        if len(images) == 0:
-            raise FileNotFoundError(f"No images could be loaded from directory '{directory}'.")
-
-        return (torch.cat(images, dim=0), torch.stack(masks, dim=0), image_count)
-    
-    @classmethod
-    def IS_CHANGED(s, directory: str, image_load_cap: int = 0, skip_first_images: int = 0, select_every_nth: int = 1):
-        directory = folder_paths.get_annotated_filepath(directory.strip())
-        if not os.path.isdir(directory):
-            return False
-        
-        dir_files = get_sorted_dir_files_from_directory(directory, skip_first_images, select_every_nth, FolderOfImages.IMG_EXTENSIONS)
-        dir_files = dir_files[:image_load_cap]
-
-        m = hashlib.sha256()
-        for filepath in dir_files:
-            m.update(calculate_file_hash(filepath).encode()) # strings must be encoded before hashing
-        return m.digest().hex()
-
-    @classmethod
-    def VALIDATE_INPUTS(s, directory: str, **kwargs):
-        directory = folder_paths.get_annotated_filepath(directory.strip())
-        if not os.path.isdir(directory):
-            return f"Directory '{directory}' cannot be found."
-        dir_files = os.listdir(directory)
-        if len(dir_files) == 0:
-            return f"No files in directory '{directory}'."
-
-        return True
-
 
 NODE_CLASS_MAPPINGS = {
     "VHS_VideoCombine": VideoCombine,
-    "VHS_LoadVideo": LoadVideo,
-    "VHS_LoadImages": LoadImagesFromDirectory,
+    "VHS_LoadVideo": LoadVideoUpload,
+    "VHS_LoadVideoPath": LoadVideoPath,
+    "VHS_LoadImages": LoadImagesFromDirectoryUpload,
+    "VHS_LoadImagesPath": LoadImagesFromDirectoryPath,
     # Latent and Image nodes
     "VHS_SplitLatents": SplitLatents,
     "VHS_SplitImages": SplitImages,
@@ -521,8 +237,10 @@ NODE_CLASS_MAPPINGS = {
 }
 NODE_DISPLAY_NAME_MAPPINGS = {
     "VHS_VideoCombine": "Video Combine 🎥🅥🅗🅢",
-    "VHS_LoadVideo": "Load Video 🎥🅥🅗🅢",
-    "VHS_LoadImages": "Load Images 🎥🅥🅗🅢",
+    "VHS_LoadVideo": "Load Video (Upload) 🎥🅥🅗🅢",
+    "VHS_LoadVideoPath": "Load Video (Path) 🎥🅥🅗🅢",
+    "VHS_LoadImages": "Load Images (Upload) 🎥🅥🅗🅢",
+    "VHS_LoadImagesPath": "Load Images (Path) 🎥🅥🅗🅢",
     # Latent and Image nodes
     "VHS_SplitLatents": "Split Latent Batch 🎥🅥🅗🅢",
     "VHS_SplitImages": "Split Image Batch 🎥🅥🅗🅢",
